pqexpress/backend/requirements.txt
# ============================================================
# PQEXPRESS Backend - Dependencias Python
# ============================================================

# Framework web
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.23.0

# Base de datos
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0

# Autenticación y seguridad
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
